pyproject.toml
[tool.poetry]
name = "skaha"
version = "0.1.0"
description = "Python Client for Skaha Container Platform in CANFAR"
authors = ["Shiny Brar <charanjotbrar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
vos = "^3.3.4"
requests = "^2.26.0"
attrs = "^21.2.0"
validators = "^0.18.2"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
